requirements.txt
cprint==1.2.2
python-dotenv==0.20.0
spotipy==2.20.0
youtube_dl==2021.12.17
pytube
